requirements.txt
Flask==2.3.2
gunicorn==20.1.0
pillow==10.0.0
firebase-admin==6.2.0
google-cloud-secret-manager==2.16.3